+++ simple-stack/src/main/java/com/zhuinden/simplestack/BackstackDelegate.java
@@ -17,6 +17,7 @@ package com.zhuinden.simplestack;
 
 import android.content.Context;
 import android.os.Bundle;
+import android.os.Parcel;
 import android.os.Parcelable;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
@@ -41,6 +42,55 @@ import java.util.Set;
  * This should be used in Activities to make sure that the {@link Backstack} survives both configuration changes and process death.
  */
 public class BackstackDelegate {
+    private static class ParcelledState
+            implements Parcelable {
+        Parcelable parcelableKey;
+        SparseArray<Parcelable> viewHierarchyState;
+        Bundle bundle;
+
+        private ParcelledState() {
+        }
+
+        protected ParcelledState(Parcel in) {
+            parcelableKey = in.readParcelable(getClass().getClassLoader());
+            // noinspection unchecked
+            viewHierarchyState = in.readSparseArray(getClass().getClassLoader());
+            boolean hasBundle = in.readByte() > 0;
+            if(hasBundle) {
+                bundle = in.readBundle(getClass().getClassLoader());
+            }
+        }
+
+        public static final Creator<ParcelledState> CREATOR = new Creator<ParcelledState>() {
+            @Override
+            public ParcelledState createFromParcel(Parcel in) {
+                return new ParcelledState(in);
+            }
+
+            @Override
+            public ParcelledState[] newArray(int size) {
+                return new ParcelledState[size];
+            }
+        };
+
+        @Override
+        public int describeContents() {
+            return 0;
+        }
+
+        @Override
+        public void writeToParcel(Parcel dest, int flags) {
+            dest.writeParcelable(parcelableKey, flags);
+            // noinspection unchecked
+            SparseArray<Object> sparseArray = (SparseArray) viewHierarchyState;
+            dest.writeSparseArray(sparseArray);
+            dest.writeByte(bundle != null ? (byte) 0x01 : 0x00);
+            if(bundle != null) {
+                dest.writeBundle(bundle);
+            }
+        }
+    }
+
     private final Backstack.CompletionListener completionListener = new Backstack.CompletionListener() {
         @Override
         public void stateChangeCompleted(@NonNull StateChange stateChange) {
@@ -53,19 +103,19 @@ public class BackstackDelegate {
         public final void handleStateChange(StateChange stateChange, Callback completionCallback) {
             //Log.i("ServiceManager", Arrays.toString(stateChange.getPreviousState().toArray()) + " :: " + Arrays.toString(stateChange.getNewState().toArray())); //
             //serviceManager.dumpLogData(); //
-            Parcelable topNewKey = stateChange.topNewState();
+            Object topNewKey = stateChange.topNewState();
             boolean isInitializeStateChange = stateChange.getPreviousState().isEmpty();
             boolean servicesUninitialized = (isInitializeStateChange && !serviceManager.hasServices(topNewKey));
             if(servicesUninitialized || !isInitializeStateChange) {
                 serviceManager.setUp(BackstackDelegate.this, topNewKey);
             }
             for(int i = stateChange.getPreviousState().size() - 1; i >= 0; i--) {
-                Parcelable previousKey = stateChange.getPreviousState().get(i);
+                Object previousKey = stateChange.getPreviousState().get(i);
                 if(serviceManager.hasServices(previousKey) && !stateChange.getNewState().contains(previousKey)) {
                     serviceManager.tearDown(BackstackDelegate.this, false, previousKey);
                 }
             }
-            Parcelable topPreviousKey = stateChange.topPreviousState();
+            Object topPreviousKey = stateChange.topPreviousState();
             if(topPreviousKey != null && stateChange.getNewState().contains(topPreviousKey)) {
                 serviceManager.tearDown(BackstackDelegate.this, true, topPreviousKey);
             }
@@ -76,6 +126,35 @@ public class BackstackDelegate {
 
     private static final String UNINITIALIZED = "";
     private String persistenceTag = UNINITIALIZED;
+
+    private KeyParceler keyParceler = new KeyParceler() {
+        @Override
+        public Parcelable toParcelable(Object object) {
+            return (Parcelable) object;
+        }
+
+        @Override
+        public Object fromParcelable(Parcelable parcelable) {
+            return parcelable;
+        }
+    };
+
+    /**
+     * Specifies a custom {@link KeyParceler}, allowing key parcellation strategies to be used for turning a key into Parcelable.
+     *
+     * If used, this method must be called before {@link BackstackDelegate#onCreate(Bundle, Object, ArrayList)}.
+     *
+     * @param keyParceler The custom {@link KeyParceler}.
+     */
+    public void setKeyParceler(KeyParceler keyParceler) {
+        if(backstack != null) {
+            throw new IllegalStateException("Custom key parceler should be set before calling `onCreate()`");
+        }
+        if(keyParceler == null) {
+            throw new IllegalArgumentException("The key parceler cannot be null!");
+        }
+        this.keyParceler = keyParceler;
+    }
 
     private static final String HISTORY = "simplestack.HISTORY";
     private static final String STATES = HISTORY + "_STATES";
@@ -113,7 +192,7 @@ public class BackstackDelegate {
 
     Backstack backstack;
 
-    Map<Parcelable, SavedState> keyStateMap = new HashMap<>();
+    Map<Object, SavedState> keyStateMap = new HashMap<>();
 
     StateChanger stateChanger;
 
@@ -260,17 +339,26 @@ public class BackstackDelegate {
      * @param nonConfigurationInstance The {@link NonConfigurationInstance} that is typically obtained with getLastCustomNonConfigurationInstance().
      * @param initialKeys              A list of the keys that are used to set as initial history of the backstack.
      */
-    public void onCreate(@Nullable Bundle savedInstanceState, @Nullable Object nonConfigurationInstance, @NonNull ArrayList<Parcelable> initialKeys) {
+    public void onCreate(@Nullable Bundle savedInstanceState, @Nullable Object nonConfigurationInstance, @NonNull ArrayList<Object> initialKeys) {
         if(nonConfigurationInstance != null && !(nonConfigurationInstance instanceof NonConfigurationInstance)) {
             throw new IllegalArgumentException(
                     "The provided non configuration instance must be of type BackstackDelegate.NonConfigurationInstance!");
         }
-        ArrayList<Parcelable> keys;
+        ArrayList<Object> keys;
         if(savedInstanceState != null) {
-            keys = savedInstanceState.getParcelableArrayList(getHistoryTag());
-            List<SavedState> savedStates = savedInstanceState.getParcelableArrayList(getStateTag());
+            List<Parcelable> parcelledKeys = savedInstanceState.getParcelableArrayList(getHistoryTag());
+            keys = new ArrayList<>();
+            if(parcelledKeys != null) {
+                for(Parcelable parcelledKey : parcelledKeys) {
+                    keys.add(keyParceler.fromParcelable(parcelledKey));
+                }
+            }
+            List<ParcelledState> savedStates = savedInstanceState.getParcelableArrayList(getStateTag());
             if(savedStates != null) {
-                for(SavedState savedState : savedStates) {
+                for(ParcelledState parcelledState : savedStates) {
+                    SavedState savedState = getSavedState(keyParceler.fromParcelable(parcelledState.parcelableKey));
+                    savedState.setViewHierarchyState(parcelledState.viewHierarchyState);
+                    savedState.setBundle(parcelledState.bundle);
                     keyStateMap.put(savedState.getKey(), savedState);
                 }
             }
@@ -346,11 +434,21 @@ public class BackstackDelegate {
      * @param outState the Bundle into which the backstack history and view states are saved.
      */
     public void onSaveInstanceState(@NonNull Bundle outState) {
-        outState.putParcelableArrayList(getHistoryTag(), HistoryBuilder.from(backstack).build());
-        List<Parcelable> history = backstack.getHistory();
-        serviceManager.persistServicesForKey(this, ServiceManager.ROOT_KEY);
-        serviceManager.persistServicesForKeyHierarchy(this, history.get(history.size() - 1));
-        outState.putParcelableArrayList(getStateTag(), new ArrayList<>(keyStateMap.values()));
+        ArrayList<Parcelable> history = new ArrayList<>();
+        for(Object key : backstack.getHistory()) {
+            history.add(keyParceler.toParcelable(key));
+        }
+        outState.putParcelableArrayList(getHistoryTag(), history);
+
+        ArrayList<ParcelledState> states = new ArrayList<>();
+        for(SavedState savedState : keyStateMap.values()) {
+            ParcelledState parcelledState = new ParcelledState();
+            parcelledState.parcelableKey = keyParceler.toParcelable(savedState.getKey());
+            parcelledState.viewHierarchyState = savedState.getViewHierarchyState();
+            parcelledState.bundle = savedState.getBundle();
+            states.add(parcelledState);
+        }
+        outState.putParcelableArrayList(getStateTag(), states);
     }
 
     /**
@@ -411,7 +509,7 @@ public class BackstackDelegate {
      */
     public void persistViewToState(@Nullable View view) {
         if(view != null) {
-            Parcelable key = ManagedContextWrapper.getKey(view.getContext());
+            Object key = ManagedContextWrapper.getKey(view.getContext());
             if(key == null) {
                 throw new IllegalArgumentException("The view [" + view + "] contained no key!");
             }
@@ -437,7 +535,7 @@ public class BackstackDelegate {
         if(view == null) {
             throw new IllegalArgumentException("You cannot restore state into null view!");
         }
-        Parcelable newKey = ManagedContextWrapper.getKey(view.getContext());
+        Object newKey = ManagedContextWrapper.getKey(view.getContext());
         SavedState savedState = getSavedState(newKey);
         view.restoreHierarchyState(savedState.getViewHierarchyState());
         if(view instanceof Bundleable) {
@@ -453,7 +551,7 @@ public class BackstackDelegate {
      * @return the saved state that belongs to the given key.
      */
     @NonNull
-    public SavedState getSavedState(@NonNull Parcelable key) {
+    public SavedState getSavedState(@NonNull Object key) {
         if(key == null) {
             throw new IllegalArgumentException("Key cannot be null!");
         }
@@ -469,25 +567,25 @@ public class BackstackDelegate {
         }
     }
 
-    private void clearStatesNotIn(@NonNull Map<Parcelable, SavedState> keyStateMap, @NonNull StateChange stateChange) {
-        Set<Parcelable> retainedKeys = new LinkedHashSet<>();
+    private void clearStatesNotIn(@NonNull Map<Object, SavedState> keyStateMap, @NonNull StateChange stateChange) {
+        Set<Object> retainedKeys = new LinkedHashSet<>();
         retainedKeys.add(ServiceManager.ROOT_KEY);
-        for(Parcelable key : stateChange.getNewState()) {
+        for(Object key : stateChange.getNewState()) {
             buildKeysToKeep(key, retainedKeys);
         }
         retainedKeys.addAll(getAdditionalRetainedKeys(stateChange));
         keyStateMap.keySet().retainAll(retainedKeys);
     }
 
-    protected Collection<? extends Parcelable> getAdditionalRetainedKeys(@NonNull StateChange stateChange) {
+    protected Collection<? extends Object> getAdditionalRetainedKeys(@NonNull StateChange stateChange) {
         return Collections.emptySet();
     }
 
-    private void buildKeysToKeep(Parcelable key, Set<Parcelable> retainedKeys) {
+    private void buildKeysToKeep(Object key, Set<Object> retainedKeys) {
         retainedKeys.add(key);
         if(key instanceof Services.Composite) {
-            List<? extends Parcelable> children = ((Services.Composite) key).keys();
-            for(Parcelable childKey : children) {
+            List<? extends Object> children = ((Services.Composite) key).keys();
+            for(Object childKey : children) {
                 buildKeysToKeep(childKey, retainedKeys);
             }
         }
@@ -503,7 +601,7 @@ public class BackstackDelegate {
      * @return the context to use used with LayoutInflater.from().
      */
     @NonNull
-    public Context createContext(Context base, Parcelable key) {
+    public Context createContext(Context base, Object key) {
         return new ManagedContextWrapper(base, key, serviceManager.findServices(key));
     }