+++ simple-stack/src/main/java/com/zhuinden/simplestack/ManagedContextWrapper.java
@@ -17,8 +17,8 @@ package com.zhuinden.simplestack;
 
 import android.content.Context;
 import android.content.ContextWrapper;
-import android.os.Parcelable;
 import android.support.annotation.NonNull;
+import android.support.annotation.Nullable;
 import android.view.LayoutInflater;
 
 /**
@@ -31,10 +31,10 @@ class ManagedContextWrapper
 
     LayoutInflater layoutInflater;
 
-    final Parcelable key;
+    final Object key;
     final Services services;
 
-    public ManagedContextWrapper(Context base, @NonNull Parcelable key, @NonNull Services services) {
+    public ManagedContextWrapper(Context base, @NonNull Object key, @Nullable Services services) {
         super(base);
         if(key == null) {
             throw new IllegalArgumentException("Key cannot be null!");
@@ -52,7 +52,7 @@ class ManagedContextWrapper
             return layoutInflater;
         } else if(TAG.equals(name)) {
             return key;
-        } else {
+        } else if(services != null) {
             Object service = services.getService(name);
             if(service != null) {
                 return service;
@@ -67,7 +67,7 @@ class ManagedContextWrapper
      * @param context the key context wrapper in which the key can be found.
      * @return the key.
      */
-    public static <T extends Parcelable> T getKey(Context context) {
+    public static <T extends Object> T getKey(Context context) {
         // noinspection ResourceType
         Object key = context.getSystemService(TAG);
         // noinspection unchecked