+++ simple-stack/src/main/java/com/zhuinden/simplestack/SavedState.java
@@ -16,10 +16,8 @@
 package com.zhuinden.simplestack;
 
 import android.os.Bundle;
-import android.os.Parcel;
 import android.os.Parcelable;
 import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
 import android.util.SparseArray;
 
 /**
@@ -28,16 +26,15 @@ import android.util.SparseArray;
  *
  * A {@link SavedState} represents the state of the view that is bound to a given key.
  */
-public class SavedState
-        implements Parcelable {
-    private Parcelable key;
+public class SavedState {
+    private Object key;
     private SparseArray<Parcelable> viewHierarchyState;
     private Bundle bundle;
 
     private SavedState() {
     }
 
-    public Parcelable getKey() {
+    public Object getKey() {
         return key;
     }
 
@@ -51,6 +48,10 @@ public class SavedState
 
     Bundle getBundle() {
         return bundle;
+    }
+
+    void setBundle(Bundle bundle) {
+        this.bundle = bundle;
     }
 
     public Bundle getViewBundle() {
@@ -71,7 +72,7 @@ public class SavedState
      * Keys are not optional.
      */
     public static class Builder {
-        private Parcelable key;
+        private Object key;
         private SparseArray<Parcelable> viewHierarchyState = new SparseArray<>();
         private Bundle bundle = new Bundle();
 
@@ -79,7 +80,7 @@ public class SavedState
             bundle.putString("___EXISTENCE_HOOK", "___EXISTENCE_HOOK");
         }
 
-        public Builder setKey(@NonNull Parcelable key) {
+        public Builder setKey(@NonNull Object key) {
             if(key == null) {
                 throw new IllegalArgumentException("Key cannot be null");
             }
@@ -97,39 +98,6 @@ public class SavedState
             savedState.bundle = bundle;
             return savedState;
         }
-    }
-
-    protected SavedState(Parcel in) {
-        key = in.readParcelable(getClass().getClassLoader());
-        // noinspection unchecked
-        viewHierarchyState = in.readSparseArray(getClass().getClassLoader());
-        bundle = in.readBundle(getClass().getClassLoader());
-    }
-
-    public static final Creator<SavedState> CREATOR = new Creator<SavedState>() {
-        @Override
-        public SavedState createFromParcel(Parcel in) {
-            return new SavedState(in);
-        }
-
-        @Override
-        public SavedState[] newArray(int size) {
-            return new SavedState[size];
-        }
-    };
-
-    @Override
-    public int describeContents() {
-        return 0;
-    }
-
-    @Override
-    public void writeToParcel(Parcel dest, int flags) {
-        dest.writeParcelable(key, flags);
-        // noinspection unchecked
-        SparseArray<Object> sparseArray = (SparseArray)viewHierarchyState;
-        dest.writeSparseArray(sparseArray);
-        dest.writeBundle(bundle);
     }
 
     @Override